+++ src/getHtml.ts
@@ -16,9 +16,8 @@ export function getResourcePath(webview: vscode.Webview, context: vscode.Extensi
 /**
  * creates the html for the csv editor
  * @param context 
- * @param initialContent 
  */
-export function createEditorHtml(webview: vscode.Webview, context: vscode.ExtensionContext, initialContent: string): string {
+export function createEditorHtml(webview: vscode.Webview, context: vscode.ExtensionContext): string {
 
 	const _getResourcePath = getResourcePath.bind(undefined, webview, context)
 
@@ -174,6 +173,11 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 										<span class="mar-left-half clickable" onclick="reRenderTable()" style="margin-left: 2em;"
 											title="Redraws the table. This can fix some measuring issues (e.g. after the font size changed)">
 											<i id="re-render-table-icon" class="fas fa-ruler-combined"></i>
+										</span>
+
+										<span id="reload-file" class="clickable" onclick="preReloadFileFromDisk()" style="margin-left: 2em;"
+											title="Reload the csv file content (from disk)">
+											<i class="fas fa-sync-alt"></i>
 										</span>
 				
 										<span id="unsaved-changes-indicator" class="hoverable unsaved-changes-indicator op-hidden tooltip is-tooltip-left" style="float: right;margin-right: 5px;"
@@ -373,21 +377,21 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 			<div class="table-action-buttons">
 
 				<div class="separated-btns">
-					<button data-test="ui.btn.addRow" class="button is-outlined" onclick="addRow()">
+					<button class="button is-outlined" onclick="addRow()">
 						<span class="icon is-small">
 							<i class="fas fa-plus"></i>
 						</span>
 						<span>Add row</span>
 					</button>
 
-					<button data-test="ui.btn.addCol"  class="button is-outlined" onclick="addColumn()">
+					<button class="button is-outlined" onclick="addColumn()">
 						<span class="icon is-small">
 							<i class="fas fa-plus"></i>
 						</span>
 						<span>Add column</span>
 					</button>
 
-					<button data-test="ui.btn.applyAndSave" class="button is-outlined mar-left" onclick="postApplyContent(true)">
+					<button class="button is-outlined mar-left" onclick="postApplyContent(true)">
 						<span class="icon is-small">
 							<i class="fas fa-save"></i>
 						</span>
@@ -398,7 +402,7 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 						</span>
 					</button>
 
-					<button data-test="ui.btn.apply" class="button is-outlined" onclick="postApplyContent(false)">
+					<button class="button is-outlined" onclick="postApplyContent(false)">
 						<span class="icon is-small">
 							<i class="fas fa-reply"></i>
 						</span>
@@ -446,7 +450,7 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 							</span>
 						</button>
 
-						<button data-test="ui.btn.showHelpModal" class="button is-outlined" onclick="toggleHelpModal(true)">
+						<button class="button is-outlined" onclick="toggleHelpModal(true)">
 							<span class="icon is-small">
 								<i class="fas fa-question"></i>
 							</span>
@@ -474,7 +478,7 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 	let helpModalHtml = ``
 	{
 		helpModalHtml = `
-		<div id="help-modal" class="modal help-modal">
+		<div id="help-modal" class="modal my-modal">
 		<div class="modal-background"></div>
 		<div class="modal-content">
 			<div class="box">
@@ -552,7 +556,7 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 	let askReadAgainModalHtml = ``
 	{
 		askReadAgainModalHtml = `
-		<div id="ask-read-again-modal" class="modal help-modal">
+		<div id="ask-read-again-modal" class="modal my-modal centered">
 		<div class="modal-background"></div>
 		<div class="modal-content">
 			<div class="box">
@@ -579,6 +583,38 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 		<button class="modal-close is-large" aria-label="close" onclick="toggleAskReadAgainModal(false)"></button>
 	</div>
 	`
+	}
+
+	let askReloadFileModalHtml = ``
+	{
+		askReloadFileModalHtml =`
+		<div id="ask-reload-file-modal" class="modal my-modal centered">
+		<div class="modal-background"></div>
+		<div class="modal-content">
+			<div class="box">
+				<h3 class="title is-3">Reload file content and discard changes</h3>
+
+				<p>
+					Are you sure you want to read the source file again? <br />
+					All changes to the table will be discarded! <br />
+					This will also update the snapshot of the file that is used for the reset data feature.
+				</p>
+
+				<div style="margin-top: 1em">
+					<button class="button is-warning" onclick="reloadFileFromDisk()">
+						<span>Reload</span>
+					</button>
+
+					<button style="margin-left: 0.5em" class="button is-outlined" onclick="toggleAskReloadFileModalDiv(false)">
+						<span>Cancel</span>
+					</button>
+				</div>
+
+			</div>
+		</div>
+		<button class="modal-close is-large" aria-label="close" onclick="toggleAskReloadFileModalDiv(false)"></button>
+	</div>
+		`
 	}
 
 	return `
@@ -620,6 +656,8 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 	${helpModalHtml}
 
 	${askReadAgainModalHtml}
+
+	${askReloadFileModalHtml}
 
 	<script>
 	var initialConfig = ${JSON.stringify(config)};